fix(observe): address P1/P2 review findings for stream 6 (#265/#263)
P1-1: Add .js extensions to relative imports in sync-manager.ts,
event-listeners.ts, sync-client.ts and their compiled JS output.
MV3 service workers require explicit extensions.

P1-2: Extract GetPageInfo and helpers (resolvePageURL, resolvePageTitle)
from analysis.go (808 LOC) into page_info.go to stay under 800 LOC limit.

P2-1: Validate tab_status in sync handler — only accept "loading" or
"complete", default to empty string for unknown values.

P2-2: Add TODO for future test file split at 1000 LOC threshold.

P2-3: Add TODO for missing OpenAPI spec fields (data_age_ms,
page_ready_for_commands, tab_status).

P2-4: Fix ResponseMetadata struct field alignment to match gofmt output.
